+++ src/components/pages/components/LinkResolver/Resolvers/PasswordReset.tsx
@@ -170,7 +170,7 @@ const PasswordReset: React.FC<PasswordResetProps> = (props) => {
     />
     <button className="btn btn-primary"
       onClick={() => {
-        dispatch(dialogActions.openDialogWithoutPayload('passwordForgotten'))
+        dispatch(dialogActions.openDialogWithoutPayload({ type: 'passwordForgotten' }))
       }}
     >
       {t(`${translationRootKey}.content.restartBtn`)}