+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 - Exporting studyAPI and reudx actions so these can be used from extension components.
 - Survey component is using isLoggedIn attribute on survey context.
 - Email validation rule during signup has been replaced to allow characters that are supported by the standard (.e.g, 'x').
+- Include the possibility to customize footer column's className.
 
 ### Fixes