+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 1.5.0 - ???
+## 2.0.0 - ???
 
 ### BREAKING CHANGES
 
@@ -36,6 +36,7 @@
 - Survey component is using isLoggedIn attribute on survey context.
 - Email validation rule during signup has been replaced to allow characters that are supported by the standard (.e.g, 'x').
 - Include the possibility to customize footer column's className.
+- Dialog config includes optional info about origin, to signalize where the dialog was opened from.
 
 ### Fixes